Apple PKCS7 - Detailed comment added.
