+++ Sources/Easy-Purchase-Receipt-Validator/AppleContainer/AppleContainerErrors.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by BJIT on 9/11/23.
+//
+
+import Foundation
+
+enum ContainerError: Error {
+    case invalid
+    case parsingIssue
+}
