+++ Sources/Easy-Purchase-Receipt-Validator/AppleContainer/SignatureData.swift
@@ -1,0 +1,8 @@
+//
+//  File.swift
+//  
+//
+//  Created by BJIT on 9/11/23.
+//
+
+import Foundation
